Set ClusterOperator status.versions, relatedObjects
Update to the ClusterOperator definition in the latest openshift/api, which
replaces the .status.version field with a .status.versions field and adds a
.status.relatedObjects field.

This commit resolves NE-144.

https://jira.coreos.com/browse/NE-144

* cmd/cluster-dns-operator/main.go (main): Set Config in Handler.
* pkg/stub/handler.go (Handler): Add Config field.
* pkg/stub/status.go (syncOperatorStatus): Set .status.versions and
.status.relatedObjects.
* pkg/util/clusteroperator/status.go (ObjectReferencesEqual): New function
to help test whether the related objects changed in the status.
(VersionsEqual): New function to help test whether the operand versions
changed in the status.
* pkg/util/clusteroperator/status_test.go (TestObjectReferencesEqual):
(TestVersionsEqual): New tests.