Set requests on operator and on daemonset
Set resources.requests.cpu=10m on the operator's container; set
resources.requests.cpu=100m, resources.requests.memory=70Mi, and
resources.limits.memory=512Mi on the CoreDNS container; and set
resources.requests.cpu=10m on the node resolver sidecar container.

The setting for the operator is the standard setting for operators that run
on masters.  The requests setting for CoreDNS is based on the upstream
default whereas the limit is based on upstream discussions, which indicate
that the upstream default limit of 170Mi may be too low at scale.  The
setting for the sidecar is based on guidance that "it's ok for them to be
low to start" and absence of data.

This commit resolves NE-141.

https://jira.coreos.com/browse/NE-141

* assets/dns/daemonset.yaml: Set resources.requests.cpu=100m,
resources.requests.memory=70Mi, and resources.limits.memory=512Mi on the
CoreDNS container.  Set resources.requests.cpu=10m on the sidecar.
* manifests/0000_08_cluster-dns-operator_02-deployment.yaml: Set
resources.requests.cpu=10m on the operator container.
* pkg/manifests/bindata.go: Regenerate.