Fix Progressing condition upgrade check
Fix-up to commit e1fefeef02373f5c2233e6670713620b5bc42b44.

This commit is related to bug 1690342.

https://bugzilla.redhat.com/show_bug.cgi?id=1690342

* pkg/operator/controller/status.go (computeOperatorProgressingCondition):
Move the check for a mismatch between old version and reported version
before the check for a mismatch between current version and reported
version.  Make sure to use the right string for the condition's message.