Report operator is degraded if any DNS is degraded
If any DNS resource's degraded condition is true, set the clusteroperator's
degraded condition to be true as well.

Note: An implication of this change is that the clusteroperator will report
degraded true during upgrades.

This commit fixes bug 1711364.

https://bugzilla.redhat.com/show_bug.cgi?id=1711364

* pkg/operator/controller/status.go (dnsStatusConditionsCounts): New struct
type.  Define a field for each status condition, to count how many DNSes
have that condition, as well as field for the total number of DNSes.
(syncOperatorStatus): Use computeDNSStatusConditionCounts to get a
dnsStatusConditionsCounts value, and pass this value to
computeOperatorStatusVersions and computeOperatorStatusConditions.  Delete
the dnsesTotal, dnsesAvailable, and allDNSesAvailable variables.
(computeDNSStatusConditionCounts): New function.  Take a slice of DNSes and
return a dnsStatusConditionsCounts value.
(computeOperatorStatusVersions): Replace the allDNSesAvailable parameter
with a dnsStatusConditionsCounts parameter.
(computeOperatorStatusConditions): Replace the dnses and dnsesAvailable
parameters with a dnsStatusConditionsCounts parameter.
(computeOperatorDegradedCondition):
(computeOperatorProgressingCondition): Replace the dnsesAvailable and
dnsesTotal parameters with a dnsStatusConditionsCounts parameter.
(computeOperatorAvailableCondition): Replace the dnsesAvailable parameter
with a dnsStatusConditionsCounts parameter.
(numDNSesAvailable): Delete.
* pkg/operator/controller/status_test.go
(TestComputeOperatorStatusConditions): Use dnsStatusConditionsCounts.
Adjust test cases to expect the degraded condition to be true during
upgrades.
(TestComputeOperatorStatusVersions): Use dnsStatusConditionsCounts.