Switch to Go modules, Go 1.12, and latest libraries
Migrate to Go modules. Switch to Go 1.12. Upgrade to Kube 1.14 and the latest
controller-runtime.

Switch CRD generation away from the openshift fork and to the upstream
controller-tools generator.

Motivations for modules:

1. Modules resolves a transitive controller-tools conflict between
controller-runtime and library-go which `dep` could not accomodate.

2. Kubernetes will soon be using modules, and some related community projects on
which we depend are also migrating. The Kubernetes KEP has a good motivation
section which applies very broadly[1].

There are some important notes for contributors:

1. Vendoring is enabled for simple reproducible builds. It's therefore important
to explicitly enable vendoring in various Go execution contexts if
reproducibility is a concern. All the build tooling has been updated to enable
vendoring. Any end-user scripts or tooling should be audited for vendoring
considerations.[2]

2. If your project lives in a GOPATH, you may have to move your project or
perform extra steps to activate modules.

[1] https://github.com/kubernetes/enhancements/blob/master/keps/sig-architecture/2019-03-19-go-modules.md#motivation
[2] https://github.com/golang/go/wiki/Modules#how-do-i-use-vendoring-with-modules-is-vendoring-going-away
[3] https://github.com/golang/go/wiki/Modules#how-to-use-modules