Integrate operator metrics with Prometheus
The operator exposes metrics, but they are not collected.  Configure
Prometheus to collect these metrics.

This commit resolves NE-76.

https://jira.coreos.com/browse/NE-76

* manifests/0000_70_dns-operator_00-cluster-role.yaml: Allow
kube-rbac-proxy to create subjectaccessreviews and tokenreviews.
* manifests/0000_70_dns-operator_00-namespace.yaml: Add a label for
openshift-monitoring.
* manifests/0000_70_dns-operator_01-service.yaml: New file.  Define a
metrics service.
* manifests/0000_70_dns-operator_02-deployment.yaml: Use kube-rbac-proxy to
expose the metrics port.
* manifests/0000_90_dns-operator_00_prometheusrole.yaml:  New file.
Define a role that grants access to services, endpoints, and pods.
* manifests/0000_90_dns-operator_01_prometheusrolebinding.yaml:  New
file.  Bind the prometheus-k8s service account to the new role.
* manifests/0000_90_dns-operator_02_servicemonitor.yaml: New file.  Define
a service monitor for the new service.
* manifests/image-references: Add kube-rbac-proxy.