Bug 1756344: Add `policy sequential` to Corefile
This instructs CoreDNS to always try the first forwarder in the list
rather then a random one (the default behaviour).

OpenStack and BareMetal rely on this with their static coredns pods.
They set the first `resolv.conf` entry to the DNS serving the
necessary OpenShift records (`api`, `api-int`, `*.apps`, etcd SRV and
hostnames) and forward anything else to all the subsequent
nameservers (see[1][2] for more details about the architecture).

[1]: https://github.com/openshift/installer/blob/master/docs/design/openstack/networking-infrastructure.md
[2]: https://github.com/openshift/installer/blob/master/docs/design/baremetal/networking-infrastructure.md

When the cluster DNS (`openshift-dns/dns-default`) pods randomly
select anything but the first entry it will fail to resolve the
records above. This happens frequently for the `default-dns`
service (`172.30.0.10`) and results in issues such as the
`authentication-operator` becoming degraded every few minutes.