Fix "unbound variable" error on Bash 4.2
Fix an empty-value check in the node-resolver shell-script to be compatible
with both Bash 4.2 and Bash 4.4.

The previous commit fixed the test to work on Bash 4.4 but caused the test
to fail with "unbound variable" on Bash 4.2 (see
https://bugzilla.redhat.com/show_bug.cgi?id=1875566).

Follow-up to commit 5c7efff2a3288271a891748ac2919417d3c6a347.

This commit fixes bug 1876912.

https://bugzilla.redhat.com/show_bug.cgi?id=1876912

* assets/dns/daemonset.yaml: Use ${svc_ips[*]-} to avoid "unbound variable"
error.
* pkg/manifests/bindata.go: Regenerate.