Fix operator namespace in clusteroperator related resources
The dns cluster operator is fed the DNS operator's namespace in
pkg/operator/controller/status/controller.go. This commit fixes a
regression brought in by https://github.com/openshift/cluster-dns-operator/pull/187
in which the cluster operator is given a reference to the "dns"
namespace, which does not exist. This commit instead gives the cluster
operator a reference to the correct operator namespace set at run
time.