status: Verify and report endpoints availability
* pkg/operator/controller/controller.go (ensureDNS): Get the DNS service's
endpoints and pass it to syncDNSStatus.
* pkg/operator/controller/dns_status.go (syncDNSStatus): Add endpoints
parameter.  Use endpoints and the new checkEndpoints function to verify
endpoints, and pass the results to computeDNSStatusConditions.
(checkEndpoints): New function.  Verify the provided endpoints by sending a
DNS query for the API service to each endpoint, and return a list of
errors.
(computeDNSStatusConditions): Add parameter for resolver errors.  Pass the
errors to computeDNSDegradedCondition.
(computeDNSDegradedCondition): Add parameter for resolver errors.  Report
Degraded=True if the list of resolver errors is nonempty,
* pkg/operator/controller/dns_status_test.go (TestDNSStatusConditions): Add
test case for resolver errors.