Configure metrics endpoints to use localhost
Configure the operator's metrics handler and the CoreDNS metrics plugin
to listen on 127.0.0.1 only.

Before this commit, metrics were exposed on ports 9153 and 60000, without
authentication.  Although the pod spec does not advertise ports 9153 or
60000, the ports were still accessible outside the pod.  After this commit,
ports 9153 and 60000 are inaccessible outside the pod, so the only way to
get metrics is by connecting to port 9154 (for CoreDNS metrics) or port
9393 (for operator metrics) and authenticating with kube-rbac-proxy.

This commit fixes bug 1896923.

https://bugzilla.redhat.com/show_bug.cgi?id=1896923

* cmd/dns-operator/main.go (main): Configure metrics to listen on
127.0.0.1:60000.
* pkg/operator/controller/controller_dns_configmap.go (corefileTemplate):
Configure metrics to listen on 127.0.0.1:9153.
* pkg/operator/controller/controller_dns_configmap_test.go
(TestDesiredDNSConfigmap): Update the expected Corefile accordingly.