Set DNS DaemonSet's maxUnavailable value to 10%
Enable rolling updates to scale linearly with the number of nodes by
setting the maxUnavailable value to a percentage instead of the default
value of 1.

This commit fixes bug 1903887.

https://bugzilla.redhat.com/show_bug.cgi?id=1903887

* assets/dns/daemonset.yaml: Set maxUnavailable to 10%.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/controller_dns_daemonset.go
(daemonsetConfigChanged): Check if the update strategy changed and update
it if it did.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDaemonsetConfigChanged): Add test case to verify that
daemonsetConfigChanged detects updates to spec.updateStrategy.
* pkg/operator/controller/dns_status.go
(getIntOrPercentValueSafely, getScaledValueFromIntOrPercent): New functions
based on definitions in Kubernetes 1.20.
(computeDNSDegradedCondition): Handle percentages values for
maxUnavailable.
* pkg/operator/controller/dns_status_test.go
(TestComputeDNSDegradedCondition): New test.