Set DNS DaemonSet's maxUnavailable value to 10%
Enable rolling updates to scale linearly with the number of nodes by
setting the maxUnavailable value to a percentage instead of the default
value of 1.

This commit fixes bug 1880148.

https://bugzilla.redhat.com/show_bug.cgi?id=1880148

* assets/dns/daemonset.yaml: Set maxUnavailable to 10%.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/controller_dns_daemonset.go
(daemonsetConfigChanged): Check if the update strategy changed and update
it if it did.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDaemonsetConfigChanged): Add test case to verify that
daemonsetConfigChanged detects updates to spec.updateStrategy.
* pkg/operator/controller/dns_status.go (computeDNSDegradedCondition):
Handle percentages values for maxUnavailable.
* pkg/operator/controller/dns_status_test.go
(TestComputeDNSDegradedCondition): New test.