Prevent dig errors from corrupting host's /etc/hosts
It's possible for `dig` to print error messages and still succeed, such as
if retries are attempted. If that happens, each word of the error
message is treated as a separate IP for the service that's being
resolved, causing /etc/hosts to be updated with garbage data.

This PR does 2 things:
- filters out error messages from `dig`, which always begin the line
  with ';' (usually with ';;', but some messages only have one
  semicolon)
- sets maximum retries to 0 when using tcp for dns, so that the behavior
  matches that of udp