Set CoreDNS's cache's maximum TTL to 900 seconds
Before this commit, the operator configured CoreDNS's cache plugin with a
maximum TTL of 30 seconds, meaning cache entries expired after 30 seconds
(or sooner, depending on the TTL from the upstream resolver).  This commit
increases the maximum TTL to 900 seconds to reduce load on upstream
resolvers.

This commit fixes bug 1933761.

https://bugzilla.redhat.com/show_bug.cgi?id=1933761

* pkg/operator/controller/controller_dns_configmap.go (corefileTemplate):
Configure the cache plugin with a 900-second maximum TTL.
* pkg/operator/controller/controller_dns_configmap_test.go
(TestDesiredDNSConfigmap): Update to expect the new value.