daemonsetConfigChanged: Fix for grace defaulting
When comparing daemonsets to determine whether an update is required, treat
implicit and explicit default values for terminationGracePeriodSeconds as
equal.

Before this commit, the update logic would keep trying to revert the
default value that the API set for the daemonset's
terminationGracePeriodSeconds field.

Follow-up to commit f094ddf7edc95dad8398179482687bc2a7a0c15b, which added
the logic which reconciles terminationGracePeriodSeconds.

This commit is related to bug 1936022.

https://bugzilla.redhat.com/show_bug.cgi?id=1936022

* pkg/operator/controller/controller_dns_daemonset.go
(daemonsetConfigChanged): Use the new cmpTerminationGracePeriodSeconds
function to compare terminationGracePeriodSeconds values so that two
daemonsets are considered equal if the only difference between them is that
one does not specify terminationGracePeriodSeconds and the other daemonset
has the default value.
(cmpTerminationGracePeriodSeconds): New function.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDaemonsetConfigChanged): Verify that daemonsetConfigChanged returns
false if the only mutation is that terminationGracePeriodSeconds has been
updated from empty to the default value.