status: Check MaxUnavailable as late as possible
Check if the DNS daemonset has 0 pods, or if the DNS service does not have
a cluster IP address, before checking if MaxUnavailable is valid.  We can
definitively report that DNS is degraded if the DNS daemonset has 0 pods or
if the DNS service has no cluster IP address, and by checking these
conditions first, we avoid reporting that that the status is unknown when
one of the first two conditions is met and MaxUnavailable is also invalid.

* pkg/operator/controller/dns_status.go (computeDNSDegradedCondition): Move
the check for an invalid value for MaxUnavailable after the checks for
having 0 pods or empty service's cluster IP address.
* pkg/operator/controller/dns_status_test.go
(TestComputeDNSDegradedCondition): Add test case to verify that
computeDNSDegradedCondition determines that the DNS is degraded if 0 pods
are available and MaxUnavailable is invalid.