Implement DNS pod placement API
This commit implements NE-547.

https://issues.redhat.com/browse/NE-547

* pkg/operator/controller/controller_dns_daemonset.go
(desiredDNSDaemonSet): Use the node selector and tolerations, if specified,
from the dns to configure the DNS daemonset.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDesiredDNSDaemonset): Verify that the default node selector and
tolerations are set.
(TestDesiredDNSDaemonsetNodePlacement): New test.  Verify that
desiredDNSDaemonSet copies the node selector and tolerations from the dns
to the daemonset.