Guard against overly restrictive node placement
When handling an update for the dns daemonset, check whether the daemonset
currently has any pods and whether the update would prevent any pods from
being scheduled, and ignore the update if both of those conditions are met
to reduce the risk that the administrator will break cluster DNS with
poorly chosen node-placement parameters.

* manifests/0000_70_dns-operator_00-cluster-role.yaml: Grant the operator
access to nodes.
* pkg/operator/controller/controller_dns_daemonset.go (updateDNSDaemonSet):
Check if the update is safe, using the new daemonsetUpdateIsSafe method,
and omit the updated node selector and tolerations from the update if they
are not safe.
(daemonsetUpdateIsSafe): New method.  Given the current and updated dns
daemonsets, check whether the current daemonset has pods and whether the
updated daemonset's node selector and tolerations would allow any pods to
be scheduled, and return a value indicating whether the update is safe.
(tolerationsTolerateTaints): New helper function for daemonsetUpdateIsSafe.
* test/e2e/operator_test.go (TestDNSNodePlacement): New test.  Verify that
the node-placement API behaves correctly when a node selector for master
nodes is configured and when a node selector that selects no nodes is
configured.