computeDNSProgressingCondition: Check node selector
Compare the node selector and tolerations that are specified in the dns
with those specified in the daemonset, and report that the operator is
progressing if the values do not match.

* pkg/operator/controller/dns_status.go (syncDNSStatus): Pass the dns to
computeDNSStatusConditions.
(computeDNSStatusConditions): Replace the conditions parameter with a
parameter for the dns.  Pass the dns to computeDNSProgressingCondition.
(computeDNSProgressingCondition): Add a parameter for the dns.  Compare the
dns's node selector and tolerations with the daemonset's, and report if
they do not match.
* pkg/operator/controller/dns_status_test.go (TestDNSStatusConditions):
Update to call computeDNSStatusConditions with an empty dns.
(TestComputeDNSProgressingCondition): New test.  Verify that
computeDNSProgressingCondition behave as expected.