status: Don't assume dns exists
When updating the clusteroperator's status, check if a dns exists at all,
and properly handle the case where it does not.

* pkg/operator/controller/status/controller.go (Reconcile): Use the new
haveNamespace field from operatorState to determine if the operand
namespace exists.  Pass the new haveDNS field as well as the dns field from
operatorState to computeOperatorStatusVersions,
computeOperatorAvailableCondition, computeOperatorProgressingCondition, and
computeOperatorDegradedCondition.
(operatorState): Add haveNamespace and haveDNS fields.  Change the type of
the namespace field from a pointer type to a non-pointer type.
(getOperatorState): Delete now-unnecessary temporary variable.  Check if at
least 1 dns exists.  Set the haveNamespace and haveDNS fields of
operatorState.
(computeOperatorStatusVersions): Replace the allDNSesAvailable parameter
with haveDNS and dns parameters.  Return the old versions if there is no
dns or the dns reports Available=False.
(computeOperatorDegradedCondition): Replace the allDNSesAvailable parameter
with haveDNS and dns parameters.  Report Degraded=True if there is no dns
or the dns reports Degraded=True.
(computeOperatorDegradedCondition): Replace the dnsAvailable parameter with
haveDNS and dns parameters.  Report Progressing=True if there is no dns or
the dns reports Available=False.
(computeOperatorAvailableCondition): Replace the dnsAvailable parameter
with haveDNS and dns parameters.  Report Available=False if there is no dns
or the dns reports Available=False.
* pkg/operator/controller/status/controller_test.go
(TestComputeOperatorProgressingCondition): Add "dns does not exist" test
case.  Update call to computeOperatorProgressingCondition to pass in a dns
when appropriate.
(TestComputeOperatorStatusVersions): Add "initialize versions, dns does not
exist" test case.  Update call to computeOperatorStatusVersions to pass in
a dns when appropriate.