Fix status reporting for default node placement
Fix the computation of the dns's "Progressing" status condition to report
correct status when the default node selector or tolerations are used.
Before this commit, the dns reported "Progressing=True" when no node
selector or no tolerations was specified.

Improve the readability of the status condition message when reporting that
the daemonset does not have the expected node selector or tolerations.

This commit fixes bug 1946929.

https://bugzilla.redhat.com/show_bug.cgi?id=1946929

* assets/dns/daemonset.yaml: Delete tolerations and add a comment that
tolerations is set at run-time.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/controller_dns_daemonset.go
(desiredDNSDaemonSet): Factor out the defaulting logic for the node
selector and tolerations from here...
(nodeSelectorForDNS, tolerationsForDNS): ...to here.  New functions for
computing the node selector and tolerations for DNS pods.
* pkg/operator/controller/dns_status.go (computeDNSProgressingCondition):
Use the new nodeSelectorForDNS and tolerationsForDNS functions to determine
what the desired node selector and tolerations are for comparison with the
actual node selector and tolerations.  Replace %v with %+v when printing
selectors and tolerations to improve readability.
* pkg/operator/controller/dns_status_test.go (TestDNSStatusConditions): Use
the new nodeSelectorForDNS and tolerationsForDNS functions.
(TestComputeDNSProgressingCondition): Update test cases to use the default
node selector and tolerations, and add explicit test cases where a custom
node selector or tolerations is specified.