serviceChanged: Fix for ipFamilies defaulting
Ignore the ipFamilies and ipFamilyPolicy fields when comparing services to
determine whether an update is required.

Before this commit, the update logic would keep trying to revert the value
that the API set for the service's ipFamilies and ipFamilyPolicy fields.

These fields are enabled by the IPv6DualStack feature gate, which is
enabled by default in Kubernetes 1.21 and OpenShift 4.8.

This commit is related to bug 1948232.

https://bugzilla.redhat.com/show_bug.cgi?id=1948232

* pkg/operator/controller/controller_dns_service.go (serviceChanged):
Ignore the IPFamilies and IPFamilyPolicy fields.
* pkg/operator/controller/controller_dns_service_test.go
(TestDNSServiceChanged): Add test case to verify that updates to
.spec.ipFamilies and .spec.ipFamilyPolicy are ignored.