Fix for daemonset maxSurge defaulting
Specify spec.updateStrategy.rollingUpdate.maxSurge: 0 in the dns and
node-resolver daemonsets.

Before this commit, the operator left the field unspecified, and as a
consequence, the update logic would keep trying to revert the default value
that the API set for this field.  The default value for this field is 0, so
specifying the value 0 has the result that the API does not need to update
the daemonsets to fill in the default value, and so the operator's update
logic is no longer triggered by the API's defaulting of this field.

In the future, we may want to set maxSurge to a positive value so as to
ensure that DNS and node-resolver pods remain available locally on each
node during rolling updates of the daemonsets.

This commit is related to bug 1948232.

https://bugzilla.redhat.com/show_bug.cgi?id=1948232

* assets/dns/daemonset.yaml: Set spec.updateStrategy.rollingUpdate.maxSurge
to 0.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/controller_dns_node_resolver_daemonset.go
(ensureNodeResolverDaemonSet): Set
spec.updateStrategy.rollingUpdate.maxSurge to 0.