status: Report old versions while progressing
Report the current status.versions value in the clusteroperator while the
clusteroperator is reporting "Progressing=True" so that the clusteroperator
doesn't report versions before it has finished upgrading to them.

This commit fixes bug 1952174.

https://bugzilla.redhat.com/show_bug.cgi?id=1952174

* pkg/operator/controller/status/controller.go (Reconcile): Compute
newVersions and pass it to computeOperatorProgressingCondition and
computeOperatorStatusVersions.  Also pass the clusteroperator's newly
computed "Progressing" status condition to computeOperatorStatusVersions.
(computeOperatorStatusVersions): Delete the haveDNS and dns parameters and
the logic that checked the dns's "Progressing" status condition.  In their
place, add a parameter for the clusteroperator's "Progressing" status
condition, and use this status condition to decide whether to return the
old versions or the updated versions.  Add a parameter for the updated
versions.
* pkg/operator/controller/status/controller_test.go
(TestComputeOperatorStatusVersions): Update unit tests.