node-resolver: Set owner reference on the dns
Configure the node-resolver daemonset with an owner reference on the dns.

The dns controller watches daemonsets in the operand namespace and enqueues
reconcile requests for a daemonset's owner, so without the owner reference,
events for the node-resolver daemonset are ignored.

This commit fixes bug 1953280.

https://bugzilla.redhat.com/show_bug.cgi?id=1953280

* pkg/operator/controller/controller.go (ensureDNS): Pass dns to
ensureNodeResolverDaemonset.
* pkg/operator/controller/controller_dns_node_resolver_daemonset.go
(ensureNodeResolverDaemonset): Add dns parameter.  Pass dns to
desiredNodeResolverDaemonSet.
(desiredNodeResolverDaemonSet): Add dns parameter.  Add an owner reference
on the dns, so that the controller's watch on daemonsets maps updates to
the node-resolver daemonset to reconcile requests for the dns.
* pkg/operator/controller/controller_dns_node_resolver_daemonset_test.go
(TestDesiredNodeResolverDaemonset): Update call to
desiredNodeResolverDaemonSet.