Corefile: Set bufsize to 512 bytes for all servers
The CoreDNS bufsize plugin was enabled with a set size
of 1232 bytes in order to ensure that DNS queries bound for
an upstream resolver contain EDNS OPT Resource Records.
This resolved Bug 1949361.

Some DNS Clients, such as Go's built-in stub DNS resolver, are not
capable of handling DNS UDP messages that are larger than 512 bytes
(even if the proper EDNS flags are included in the message).
See https://github.com/golang/go/issues/13561 &
https://github.com/golang/go/issues/6464 accordingly.

This commit configures the bufsize plugin with a reduced bufsize of 512
bytes for all servers to ensure compatibility with primitive DNS
clients.

pkg/operator/controller/controller_dns_configmap.go:

Set bufsize to 512 bytes for all servers.

pkg/operator/controller/controller_dns_configmap_test.go:

Update unit test cases.

This commit is in support of Bug 196616 and should not cause
any regression from the changes merged as a part of Bug 1949361.