serviceChanged: Fix internalTrafficPolicy
When comparing services to determine whether an update is required, treat
the empty value and default value for spec.internalTrafficPolicy as equal.

Before this commit, the update logic would keep trying to revert the
default value that the API set for the service's internal traffic policy.

The spec.internalTrafficPolicy field is newly enabled in Kubernetes
1.22 (OpenShift 4.9).

This commit fixes bug 2002461.

https://bugzilla.redhat.com/show_bug.cgi?id=2002461

* pkg/operator/controller/controller_dns_service.go (serviceChanged): Use
the new cmpServiceInternalTrafficPolicyType function so that two services
are considered equal if the only difference between them is that one does
not specify internal traffic policy and the other service has the default
value.
(cmpServiceInternalTrafficPolicyType): New function.
* pkg/operator/controller/controller_dns_service_test.go
(TestDNSServiceChanged): Verify that serviceChanged returns false if the
only mutation is that the internal traffic policy has been updated from
empty to the default value.