Enable metrics for custom upstream resolvers
Before this commit, the operator did not enable the "prometheus" plugin in
server blocks for custom upstream resolvers.  As a consequence, CoreDNS did
not report metrics for upstream resolvers and only reported metrics for the
default server block.  This commit enables the "prometheus" plugin for all
server blocks.

This commit fixes bug 2020489.

https://bugzilla.redhat.com/show_bug.cgi?id=2020489

* pkg/operator/controller/controller_dns_configmap.go (corefileTemplate):
Add the "prometheus" plugin to server blocks for custom upstream resolvers.
* pkg/operator/controller/controller_dns_configmap_test.go
(TestDesiredDNSConfigmap): Verify that desiredDNSConfigMap enables the
"prometheus" plugin for all server blocks.