desiredDNSDaemonSet: Allow autoscaler to evict
Annotate DNS pods to enable the cluster autoscaler to evict them when
scaling down a node.

Before this commit, node scale-in could cause DNS disruption for tainted
nodes.  If a node is tainted such that it has no DNS pod, then DNS queries
originating on that node are forwarded to DNS pods on other nodes.  If one
of these nodes was removed by scale-in or otherwise, the DNS pod on that
node would be abruptly removed when the node shut down, and it would take
some time for the node lifecycle controller to mark the pod as not ready,
then for the endpoints controller and endpointslice controller to remove
the pod's endpoint, and then for the service proxy to stop forwarding
queries to the pod.  This delay could cause DNS lookup failures when
queries were sent to a pod that had been on a removed node.  After this
commit, the cluster autoscaler should evict the DNS pod before removing the
node, which should prevent disruption to the DNS service.

This commit is related to bug 2061244.

https://bugzilla.redhat.com/show_bug.cgi?id=2061244

* pkg/operator/controller/controller_dns_daemonset.go
(enableDaemonSetEvictionAnnotationKey): New const for the
"cluster-autoscaler.kubernetes.io/enable-ds-eviction" pod annotation key.
(managedDNSDaemonSetAnnotations): New variable for annotation keys of
annotations that the operator manages on DNS daemonsets.  For now, this
only includes enableDaemonSetEvictionAnnotationKey.
(desiredDNSDaemonSet): Specify the
"cluster-autoscaler.kubernetes.io/enable-ds-eviction" pod annotation.
(daemonsetConfigChanged): Check whether any managed annotations changed,
and update annotations as appropriate.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDesiredDNSDaemonset): Verify that desiredDNSDaemonSet sets the
expected pod annotations and labels.
(TestDaemonsetConfigChanged): Verify that daemonsetConfigChanged returns
true if the "cluster-autoscaler.kubernetes.io/enable-ds-eviction" pod
annotation is changed and false if only some other annotation changed.