topology aware hints to prefer to keep dns traffic within a zone
Currently DNS traffic in openshift can go to any random pod on any host in any zone. Traffic should at a minimum be prefered to stay within a zone to prevent latency and reduce failure points in a given dns request. This requiresOpenshift 4.11 or TopologyAwareHints feature gate to be enabled.