Move client pods to privileged namespaces
The client pods were landing in the default namespace so in lieu of
relabeling that namespace, we're moving them to namespaces that allow
privileged pods. The client pod for TestDNSForwarding will launch in
openshift-dns and the client pod for TestDNSOverTLSForwarding will
launch in a privileged namespace created as part of the e2e test.