Set DNS DaemonSet's maxSurge value to 10%
Enable surge for DNS pods so that a rolling update of DNS pods creates new
pods *before* deleting old pods.

This change reduces the risk of disruption to DNS during upgrades.  Without
this change, a rolling update of DNS pods deletes a pod on a node first and
then create a new pod on that node, meaning that a node that had a local
DNS pod before an upgrade would temporarily not have one at some point
during the upgrade.

Always having a local DNS pod reduces the risk of disruption, especially
when using a CNI provider that supports topology-aware hints or when using
openshift-sdn and ovn-kubernetes, which have logic similar to
topology-aware hints to prefer any available local DNS pod rather than
forwarding queries to DNS pods on other nodes.

This commit also sets minReadySeconds so that the old pod only gets deleted
once the new pod has had time to start up.

This commit additionally maxUnavailable to zero because maxUnavailable may
not be non-zero when maxSurge is non-zero.

This commit fixes OCPBUGS-5943.

https://issues.redhat.com/browse/OCPBUGS-5943

* assets/dns/daemonset.yaml: Set minReadySeconds to 9, maxSurge to 10%, and
maxUnavailable to 0.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/controller_dns_daemonset.go
(daemonsetConfigChanged): Check minReadySeconds and set it in the updated
daemonset.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDesiredDNSDaemonset): Verify that minReadySeconds and the update
strategy are set correctly.
(TestDaemonsetConfigChanged): Add test cases for minReadySeconds and
maxSurge.